+++ src/libAtomVM/term.h
@@ -128,7 +128,10 @@ extern "C" {
 #define TERM_BOXED_REFC_BINARY_SIZE 6
 #define TERM_BOXED_BIN_MATCH_STATE_SIZE 4
 #define TERM_BOXED_SUB_BINARY_SIZE 4
-#define TERM_BOXED_RESOURCE_SIZE TERM_BOXED_REFC_BINARY_SIZE
+#define TERM_BOXED_REFERENCE_RESOURCE_SIZE 4
+#define TERM_BOXED_REFERENCE_RESOURCE_HEADER (((TERM_BOXED_REFERENCE_RESOURCE_SIZE - 1) << 6) | TERM_BOXED_REF)
+#define TERM_BOXED_RESOURCE_SIZE TERM_BOXED_REFERENCE_RESOURCE_SIZE
+
 #if TERM_BYTES == 8
 #define REFC_BINARY_MIN 64
 #define SUB_BINARY_MIN 16
@@ -166,6 +169,7 @@ extern "C" {
 #define TUPLE_SIZE(elems) ((int) (elems + 1))
 #define CONS_SIZE 2
 #define REFC_BINARY_CONS_OFFSET 4
+#define REFERENCE_RESOURCE_CONS_OFFSET 2
 #define LIST_SIZE(num_elements, element_size) ((num_elements) * ((element_size) + CONS_SIZE))
 #define TERM_STRING_SIZE(length) (2 * (length))
 #define TERM_MAP_SIZE(num_elements) (3 + 2 * (num_elements))
@@ -205,6 +209,9 @@ extern "C" {
 // Local ref is at most 30 bytes:
 // 2^32-1 = 4294967295 (10 chars)
 // "#Ref<0." "." ">\0" (10 chars)
+// Resource ref is at most 52 bytes:
+// 2^32-1 = 4294967295 (10 chars)
+// "#Ref<0." "." "." "." ">\0" (12 chars)
 // External ref is at most 70 bytes:
 // 2^26-1 = 67108863 (8 chars) (node, atom index)
 // 2^32-1 = 4294967295 (10 chars)
@@ -499,6 +506,25 @@ static inline bool term_is_refc_binary(term t)
         const term *boxed_value = term_to_const_term_ptr(t);
         int masked_value = boxed_value[0] & TERM_BOXED_TAG_MASK;
         return masked_value == TERM_BOXED_REFC_BINARY;
+    }
+
+    return false;
+}
+
+/**
+ * @brief Checks if a term is a reference resource
+ *
+ * @details Returns \c true if a term is a ref-counted reference resource,
+ * otherwise \c false.
+ * @param t the term that will be checked.
+ * @return \c true if check succeeds, \c false otherwise.
+ */
+static inline bool term_is_resource_reference(term t)
+{
+    /* boxed: 10 */
+    if (term_is_boxed(t)) {
+        const term *boxed_value = term_to_const_term_ptr(t);
+        return boxed_value[0] == TERM_BOXED_REFERENCE_RESOURCE_HEADER;
     }
 
     return false;
@@ -1623,6 +1649,20 @@ static inline struct RefcBinary *term_refc_binary_ptr(term refc_binary)
 }
 
 /**
+ * @brief Get the pointer off heap refc binary for a resource.
+ *
+ * @details Returns address
+ * @return offset (in words).
+ */
+static inline struct RefcBinary *term_resource_refc_binary_ptr(term resource)
+{
+    TERM_DEBUG_ASSERT(term_is_resource_reference(resource));
+
+    term *boxed_value = term_to_term_ptr(resource);
+    return (struct RefcBinary *) boxed_value[1];
+}
+
+/**
  * @brief Gets binary data
  *
  * @details Returns a pointer to stored binary data.
@@ -1867,8 +1907,8 @@ static inline term term_from_ref_ticks(uint64_t ref_ticks, Heap *heap)
     boxed_value[1] = (term) ref_ticks;
 
 #elif TERM_BYTES == 4
-    boxed_value[1] = (ref_ticks >> 4);
-    boxed_value[2] = (ref_ticks & 0xFFFFFFFF);
+    boxed_value[1] = (uint32_t) (ref_ticks >> 32);
+    boxed_value[2] = (uint32_t) ref_ticks;
 
 #else
 #error "terms must be either 32 or 64 bit wide"
@@ -1887,7 +1927,7 @@ static inline uint64_t term_to_ref_ticks(term rt)
     return boxed_value[1];
 
 #elif TERM_BYTES == 4
-    return (boxed_value[1] << 4) | boxed_value[2];
+    return ((uint64_t) boxed_value[1] << 32) | (uint64_t) boxed_value[2];
 
 #else
 #error "terms must be either 32 or 64 bit wide"
@@ -2661,7 +2701,7 @@ static inline term term_get_sub_binary_ref(term t)
 /**
  * @brief Create a resource on the heap.
  * @details This function creates a resource (obtained from `enif_alloc_resource`)
- * on the heap which must have `TERM_BOXED_RESOURCE_SIZE` free terms.
+ * on the heap which must have `TERM_BOXED_REFERENCE_RESOURCE_SIZE` free terms.
  *
  * This function does increment the reference counter as the resource is
  * added to the heap's mso list.
@@ -2672,20 +2712,29 @@ static inline term term_get_sub_binary_ref(term t)
  */
 static inline term term_from_resource(void *resource, Heap *heap)
 {
-    // Resources are currently refc binaries with a size of 0 but may be
-    // references in the future.
+    // Resources are references
     struct RefcBinary *refc = refc_binary_from_data(resource);
-    term *boxed_value = memory_heap_alloc(heap, TERM_BOXED_REFC_BINARY_SIZE);
-    boxed_value[0] = ((TERM_BOXED_REFC_BINARY_SIZE - 1) << 6) | TERM_BOXED_REFC_BINARY;
-    boxed_value[1] = (term) 0; // binary size, this is pre ERTS 9.0 (OTP-20.0) behavior
-    boxed_value[2] = (term) RefcNoFlags;
-    term ret = ((term) boxed_value) | TERM_PRIMARY_BOXED;
-    boxed_value[3] = (term) refc;
+    term *boxed_value = memory_heap_alloc(heap, TERM_BOXED_REFERENCE_RESOURCE_SIZE);
+    boxed_value[0] = TERM_BOXED_REFERENCE_RESOURCE_HEADER;
+    boxed_value[1] = (term) refc;
     // Add the resource to the mso list
     refc->ref_count++;
-    heap->root->mso_list = term_list_init_prepend(boxed_value + 4, ret, heap->root->mso_list);
+    term ret = ((term) boxed_value) | TERM_PRIMARY_BOXED;
+    heap->root->mso_list = term_list_init_prepend(boxed_value + REFERENCE_RESOURCE_CONS_OFFSET, ret, heap->root->mso_list);
     return ret;
 }
+
+/**
+ * @brief Get a resource term from a resource type and a serialization reference
+ * number.
+ *
+ * @param resource_type_ptr pointer to the resource type
+ * @param resource_serialize_ref unique reference for serialization
+ * @param heap the heap to create the resource term in
+ * @param glb the global context
+ * @return a resource term for this resource or a stale reference if the resource cannot be found
+ */
+term term_from_resource_type_and_serialize_ref(uint64_t resource_type_ptr, uint64_t resource_serialize_ref, Heap *heap, GlobalContext *glb);
 
 #ifdef __cplusplus
 }